requirements.txt
schemachange==4.0.1
snowflake-connector-python>=3.10
PyYAML>=6.0
toml>=0.10.0
